Moved functions from build_dmrpp to build_dmrpp_utils for tests
